Fixed sprite ID for ItemBlockJSON, Fixed icon returning wrong value
Might not be a proper fix but it works. Main issue is that MC reloads the texture several times during load. During this sprite ID may be set before JSON is loaded. Thus causing a 0 to be set incorrectly for the sprite ID. The fix simply changes it so the sprite ID resets each time the resources reaload.

Additionally I fixed icons returning the wrong value. This was being caused by getIcon() method not using the getIconStateKeys(). Though I suspect this was intentional and the bug is likely  there were extra states for the method than needed. Either way it works and can be overhauled later for a better fix.